Working Adjustable Gap and Border Settings
User can now dynamically adjust their tag settings, but these settings are not saved to the JSON file yet.  
Will continue to add adjustable features and then worry about having it save.  
Border size may be too much considering that the gap works in a similar way.  
Try not to assume what the user wants, give them the freedom to edit how they want, but to an extent.