Height and Width Adjustable Inputs
User is now able to set the width, height, ratio, and the border radius of an item.
Next I will need to make a separate border radius for the floating border and then a checkbox that says to match existing border.  Also need a button that sets the size and gap of the floating border to be automatically sized and centered.